ArchiveController class now successfully reads in all tickets as ticket object through getAllTickets(), then archivesOldTickets() sets the status of closed and (un)resolved tickets if its been closed for longer than 24hrs.
