--- OpenTicket
@@ -1,10 +1,0 @@
-5,9001,4234,no post,low,open,2023-09-27T19:07:32.256613139,null
-6,9001,4000,no hdmi,low,open,2023-09-27T19:09:59.743468495,null
-7,9000,4234,no power,high,open,2023-09-27T19:11:38.291423782,null
-7,9002,4234,No POST no display,open,medium,2023-09-27T19:11:52.132140990,null
-8,9001,4000,Fan error,low,open,2023-09-27T19:12:20.626268450,null
-9,9000,4172,BSOD on boot,high,open,2023-09-27T19:13:02.874609993,null
-13,9000,4000,no power,high,open,2023-09-27T23:14:23.968558425,null
-14,9003,4000,No power,high,open,2023-10-02T13:28:11.985318295,null
-15,9004,4000,no display on dock,high,open,2023-10-02T13:28:35.203273376,null
-16,9002,4006,Its cooked,Bad,open,2023-10-06T16:53:56.330222800,null
