+++ records/OpenTicket.csv
@@ -7,5 +7,5 @@
 13,9000,4000,no power,high,open and unresolved,2023-09-27T23:14:23.968558425,null
 14,9003,4000,No power,high,open and unresolved,2023-10-02T13:28:11.985318295,null
 15,9004,4000,no display on dock,high,open and unresolved,2023-10-02T13:28:35.203273376,null
-16,9002,4006,Its cooked,Bad,closed and unresolved,2023-10-06T16:53:56.330222800,2023-10-11T17:11:52.840573500
-17,9002,4006,123,123,closed and unresolved,2023-10-07T17:11:52.840573500,null
+16,9002,4006,Its cooked,Bad,closed and unresolved,2023-10-06T16:53:56.330222800,2023-10-13T19:44:04.276697
+17,9002,4006,123,123,closed and unresolved,2023-10-07T17:11:52.840573500,2023-10-13T20:01:31.299275500
